requirements.txt
python-dotenv
openai>=1.40.0
faiss-cpu
numpy
tqdm
pypdf
python-docx
markdown
streamlit
